+++ src/QueryBuilder/EagerLoading/EagerLoadIterator.php
@@ -17,7 +17,7 @@
 
 declare(strict_types=1);
 
-namespace LaravelJsonApi\Eloquent\EagerLoading;
+namespace LaravelJsonApi\Eloquent\QueryBuilder\EagerLoading;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;