+++ tests/lib/Acceptance/EagerLoading/EagerLoaderTest.php
@@ -22,7 +22,7 @@ namespace LaravelJsonApi\Eloquent\Tests\Acceptance\EagerLoading;
 use App\Models\Post;
 use App\Models\User;
 use LaravelJsonApi\Core\Query\IncludePaths;
-use LaravelJsonApi\Eloquent\EagerLoading\EagerLoader;
+use LaravelJsonApi\Eloquent\QueryBuilder\EagerLoading\EagerLoader;
 use LaravelJsonApi\Eloquent\Tests\Acceptance\TestCase;
 
 class EagerLoaderTest extends TestCase